+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Changed
 - Adjust how fan graph in ROGCC works, deny incorrect graphs
+- Fix to apply the fan curve change in ROGCC to the correct profile
 - Support for G713RS LED modes (Author: Peter Ivanov)
 - Support for G713RM LED modes (Author: maxbachmann)
 - Fix VivoBook detection